backend/requirements.txt
flask
python-dotenv
pymongo
flask_jwt_extended
flask_restful
requests
passlib
Flask-APScheduler